front: make ownership list and owner form into separate components
